Add weather icon and script
Currently, the weather script fetches weather from wttr.in and caches it in
the .cache/dwmBlocksWeather directory. The script reruns every 60
seconds, and checks the modification time of the cache file. If it has
been 30 minutes since last cache, the script fetches a new weather
update.

Currently, the icon displays the current conditions and temp, and
clicking the icon shows more info about the current conditions in a
notification