Correct emoji output, add connectivity check
This commit corrects an issue where the trailing Variation Selector
after certain emojis would show up as a "tofu" block. The weather script
now removes Variation Selector blocks.

This commit also add a connectivity check. This script now pings
wttr.in, and if it does not recieves a response, exits. Only if the ping
recieves a response will it make a request and write to the cache file.
This prevents the script from generating a blank cache file if it tries
to run when disconnected from the internet